organise the code to match GetCtReferenceConnector and GetCtReferenceHttpParser
