+++ app/uk/gov/hmrc/incorporatedentityidentification/services/RegisterWithMultipleIdentifiersService.scala
@@ -18,7 +18,7 @@ package uk.gov.hmrc.incorporatedentityidentification.services
 
 import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.incorporatedentityidentification.connectors.RegisterWithMultipleIdentifiersConnector
-import uk.gov.hmrc.incorporatedentityidentification.connectors.RegisterWithMultipleIdentifiersHttpParser.RegisterWithMultipleIdentifiersResult
+import uk.gov.hmrc.incorporatedentityidentification.httpparsers.RegisterWithMultipleIdentifiersHttpParser.RegisterWithMultipleIdentifiersResult
 
 import javax.inject.{Inject, Singleton}
 import scala.concurrent.Future
